cleaned up code for tournament edit page
Also fixed inputs of registration on tournament view page
